builtin/k8s: Implement `WatchTask` for k8s task plugin
This commit introduces an implementation of WatchTask for the kubernetes
Task plugin. It will poll initially to ensure that the pod being watched
properly starts up before attempting to stream its logs. Once the pod
has started, WatchTask will stream the pod logs and send it to the
runner logs as well as the UI output for the job.